refactor: flatten project structure
