+++ .gitignore
@@ -1,3 +1,4 @@
 # Неотслеживаемые файлы: 
 __pycache__
 *.bkp
+*.db
