--- OrderStatus_cls.py
@@ -1,7 +1,0 @@
-"""Класс статуса заказа"""
-import enum
-
-class OrderStatus(enum.StrEnum):
-    NEW = 'Новый'
-    PREPARING = 'Готовится'
-    READY = 'Готов'
