--- restaraunts/src/restaraunts/config.py
@@ -1,6 +1,0 @@
-from pathlib import Path
-
-
-BASE_DIR = Path(__file__).resolve().parent
-
-DB_PATH = BASE_DIR / "data" / "my_database.db"
