+++ src/restaraunts/config.py
@@ -1,0 +1,7 @@
+import os
+from pathlib import Path
+
+
+BASE_DIR = Path(__file__).resolve().parent.parent.parent
+
+DB_PATH = BASE_DIR / "data" / "restaraunts.db"
