+++ src/restaraunts/main.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI
-from .endpoints import v1_router
+from src.restaraunts.endpoints import v1_router
 
 app = FastAPI()
 app.include_router(v1_router, prefix="/api")
