+++ src/restaraunts/schemas/item.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from .base import Base
+from src.restaraunts.schemas.base import Base
 
 class Item(Base):
     price: int
