+++ src/restaraunts/schemas/restaraunt.py
@@ -1,6 +1,6 @@
-from .base import Base
-from .menu import Menu
-from .order import Order
+from src.restaraunts.schemas.base import Base
+from src.restaraunts.schemas.menu import Menu
+from src.restaraunts.schemas.order import Order
 from typing import List, Optional
 
 class Restaraunt(Base):
